fix(workflows): survive failed passive refetch, reject stale reload data
TanStack Query v5 reports isError after a failed background refetch while
cached data remains. The editor host now treats a detail-query error as
fatal only when no data exists; with cached data the mounted editor and
unsaved draft survive and a non-destructive refresh warning shows. The
deliberate Reload path checks the refetch result and throws instead of
adopting the stale cached value, keeping the current draft and reporting
the failure. PersistedWorkflowEditor moves to its own file to stay under
the shared-surface line threshold.
